Added initialization code, some checks to check for initialization.
If the database has no tables, the mod will attempt to initialize
it using user-defined mc_db.sql file that should be placed in
the mod directory (DBServerStats). Each line in that file should be
blank or a query. A query must be entirely on a single line. It is
expected to have CREATE TABLE queries in that file.